Replace Arc.Ecto.Schema with importable Arc.Ecto.Changeset.
Remove the need for cast_attachments/4 to be a macro.
