+++ .gitignore
@@ -4,3 +4,4 @@ erl_crash.dump
 *.ez
 .DS_Store
 /doc
+.elixir_ls
