+++ lib/arc_ecto/changeset/helpers.ex
@@ -1,0 +1,19 @@
+defmodule Arc.Ecto.Changeset.Helpers do
+  @moduledoc false
+
+  @doc false
+  def convert_params_to_binary(params) do
+    Enum.reduce(params, nil, fn
+      {key, _value}, nil when is_binary(key) ->
+        nil
+
+      {key, _value}, _ when is_binary(key) ->
+        raise ArgumentError, "expected params to be a map with atoms or string keys, " <>
+                             "got a map with mixed keys: #{inspect params}"
+
+      {key, value}, acc when is_atom(key) ->
+        Map.put(acc || %{}, Atom.to_string(key), value)
+
+    end) || params
+  end
+end
