+++ test/changeset_test.exs
@@ -1,4 +1,4 @@
-defmodule ArcTest.Ecto.Schema do
+defmodule ArcTest.Ecto.Changeset do
   use ExUnit.Case, async: false
   import Mock
   import ExUnit.CaptureLog
@@ -6,7 +6,7 @@ defmodule ArcTest.Ecto.Schema do
   defmodule TestUser do
     use Ecto.Schema
     import Ecto.Changeset
-    use Arc.Ecto.Schema
+    import Arc.Ecto.Changeset
 
     schema "users" do
       field :first_name, :string
